requirements/base.txt
Django==6.0.4
djangorestframework==3.17.1
python-dotenv==1.1.0
